Prevent Travis from running twice on a PR
Currently, when we open a PR, Travis run twice: once because it's a
PR and once for the last commit of the PR.

This change will make Travis run for *commits* only when
made directly on master (which should not happen that much, but still
happen).
Of course, Travis will still be run for a PR, but only once.